ghost-editor/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
llama-index==0.11.0
llama-index-llms-ollama==0.3.0
llama-index-embeddings-ollama==0.3.0
llama-index-vector-stores-chroma==0.2.0
chromadb==0.5.0
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv
sqlalchemy
PyGithub
httpx
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv
pytest==8.3.3
pytest-asyncio==0.24.0
